Issue #1: Refactor move cardHasPrLinked() to utils
